Move miscellanious methods to their own classes.
